% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleData-class.R
\name{alleleData}
\alias{alleleData}
\title{alleleData}
\usage{
alleleData(data, tree, siteInfo = NULL, logProb = FALSE)
}
\arguments{
\item{data}{A list of matricies, where each element in the list is a species, each column in a matrix is an allele, and each row is a site}

\item{tree}{The tree corresponding the the list of allele data passed in. Tip names must match list names.}

\item{siteInfo}{A data.frame containing classification information about each site, where each column characteristic, and each site is a row. Defaults to assuming all sites have the same info.}

\item{logProb}{TRUE/FALSE indicating whether the passed probabilities are in log space (default=FALSE)}
}
\value{
an alleleData object
}
\description{
Contructs an object that holds the allele data and the phylogenetic tree
}
\examples{

}
