% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic}
\alias{bic}
\title{Compute BIC for rate models}
\arguments{
\item{model}{rateModel object}
}
\value{
numeric BIC value
}
\description{
Computes the bayesian information criterion (BIC) for a given rateModel.
Note that when comparing two models, the one with the smaller BIC is preferred.
}
\examples{

}
