% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disCharToProb.R
\name{disCharToProb}
\alias{disCharToProb}
\title{Discrete characters to probabailites}
\usage{
disCharToProb(x, charLevels)
}
\arguments{
\item{x}{either a matrix or a data.frame of allele values with each column being a species and each row being a site}

\item{levels}{a vector of all possible values of the discrete allele label}
}
\value{
a list of allele probabilities with one matrix per species
}
\description{
Converts discrete characters labels at each site to probabailites
}
