% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fits rate model}
\arguments{
\item{obj}{rateModel}

\item{scale}{a scale factor to apply to log-likelihood, defaults to -1/nsites}

\item{method}{Optimization method to use ("l-bfgs-b","mlsl","stogo")}

\item{threads}{number of threads to use}

\item{control}{See control from \link[stats]{optim} if using l-bfgs-b, otherwise look at \link[nloptr]{nl.opts}}
}
\value{
a list incuding information about the optimization, the model object is updated directly (by reference)
}
\description{
Fits rate model object and returns fitted model
}
\examples{

}
