% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceParameterZero.R
\name{forceParameterZero}
\alias{forceParameterZero}
\title{Remove parameter from model}
\arguments{
\item{obj}{rateModel object}

\item{index}{a 0-based integer vector of parameter indicies to be changed (note: parameters are 0 indexed)}

\item{verbose}{if TRUE prints out information about variable being frozen}
}
\description{
Sets a parameter to 0 and sets its value to be fixed so it will not be fit
}
