% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlleleData.R
\name{getAlleleData}
\alias{getAlleleData}
\title{Returns alleleData}
\arguments{
\item{model}{rateModel}
}
\value{
alleleData object
}
\description{
Gets the alleleData from a rate model object
}
\examples{

}
