% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgeTable.R
\name{getEdgeTable}
\alias{getEdgeTable}
\alias{getEdgeTable,alleleData,alleleData-method}
\alias{getEdgeTable,rateModel,rateModel-method}
\title{Get edge table}
\arguments{
\item{x}{alleleData or rateModel object}
}
\value{
data.table containing edge ids
}
\description{
Returns edge table.
}
