% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getParams}
\alias{getParams}
\alias{getParams,rateModel,rateModel-method}
\title{Plot coefficents}
\arguments{
\item{x}{rateModel object}
}
\value{
vector of parameter values
}
\description{
Returns a vector of parameter values
}
