% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateBounds.R
\name{getRateBounds}
\alias{getRateBounds}
\alias{getRateBounds,rateModel,rateModel-method}
\title{Get rate bounds}
\arguments{
\item{model}{rateModel object}
}
\value{
vector of parameter values
}
\description{
Get vector of min and max rates
}
