% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiblings.R
\name{getSiblings}
\alias{getSiblings}
\title{Get siblings}
\usage{
getSiblings(tr, node)
}
\arguments{
\item{tr}{phylo object}

\item{node}{the node number}
}
\value{
The node numbers of sibbling nodes
}
\description{
Get the siblings of a given node on a tree
}
