% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTree.R
\name{getTree}
\alias{getTree}
\title{Gets tree}
\arguments{
\item{model}{alleleData or rateModel object}
}
\value{
A tree of class \code{phylo}
}
\description{
Returns phylo object
}
\examples{

}
