% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logSumExp}
\alias{logSumExp}
\title{Vector logSumExp}
\usage{
logSumExp(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
This function computes the sum(e^x) of a vector x without leaving log space
}
