% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{lrt}
\alias{lrt}
\title{Compute LRT for rate models}
\arguments{
\item{h0}{rateModel object that represents the null hypothesis}

\item{hA}{rateModel object that represents the alternative hypothesis (more complex model)}
}
\value{
a data table including D-statistic, the p-value, and the degrees of freedom
}
\description{
Performs a likelihood ratio test for two models
}
\examples{

}
