% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixToStlXptr}
\alias{matrixToStlXptr}
\title{matrixToStlPointer}
\usage{
matrixToStlXptr(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\description{
This function converts a matrix to a 2D double stl vector and returns a XPtr to a 
pointer to the vector
}
