% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{Pack rateModel}
\arguments{
\item{model}{a rateModel object}
}
\value{
a packedModel object
}
\description{
Packs a rateModel into an S4 object that can be saved
}
\examples{

}
