% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{packedModel-class}
\alias{packedModel-class}
\title{Class packedModel}
\description{
Class \code{packedModel} holds all the data necessary to construct a rate model
}
\section{Slots}{

\describe{
\item{\code{alleleData}}{an allele data object in a locked environment}

\item{\code{edgeGroups}}{a data.table with four columns: parent, child, edgeGroup. All parent-child combos must be valid for alleleData object.}

\item{\code{rateFormula}}{A formula that uses the variables in the allele data object to compute turnover rate}

\item{\code{piFormula}}{A formula that uses the variables in the allele data object to compute pi}

\item{\code{rateDM}}{the design matrix for the rate}

\item{\code{piDM}}{the design matrix for pi}

\item{\code{phylogeny}}{parameter object holding all the parameters for phylogenetic computations}

\item{\code{fixed}}{a logical vector indicating which variables are fixed}
}}

