% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloGrad.R
\name{phyloGrad}
\alias{phyloGrad}
\title{Compute parameter gradients}
\arguments{
\item{x}{a set of parameters around which to compute the gradient (optional)}

\item{model}{rateModel object}

\item{scale}{value to scale log-likelihood by (default=1)}

\item{threads}{number of threads to use}
}
\value{
vector of gradients
}
\description{
Compute parameter gradients for given rate model object
}
