% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi}
\alias{pi}
\title{Compute site-wise stationary distribution}
\arguments{
\item{model}{rateModel object}

\item{sites}{integer vector of site indicies}
}
\value{
table of allelic probabilities along with the corresponding allele and site
}
\description{
Compute the site-wise allelic stationary distribution across specified sites
}
