% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElementTransitions.R
\name{plotElementTransitions}
\alias{plotElementTransitions}
\alias{plotElementTransitions,rateModel,rateModel-method}
\title{Plot turnover events on the tree}
\arguments{
\item{model}{rateModel}

\item{aggregate}{the type of aggregation to perform (either "edge" or "node", see \code{marignalTransitions})}

\item{excludeNodes}{Indicates nodes that should be excluded from the per-edge aggregated statistic
(1-based).}

\item{excludeEdges}{Indicates edges that should be excluded from the per-node aggregated statistic
(edge ids based on integer id of child, use plotTree(model)
to view).}

\item{offset}{How much to offset tip labels by}

\item{xmax}{Maximum value of x axis in tree plots (Note: if you want different x axis for different tree
plots use \link[ggtree]{xlim_expand} on the returned object)}

\item{nodeLabels}{if TRUE label nodes with numbers}

\item{tipLabels}{if TRUE label tips with names}
}
\description{
Plots tree with various helpful highlightings
}
\examples{

}
