% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParams.R
\name{plotParams}
\alias{plotParams}
\alias{plotParams,rateModel,rateModel-method}
\title{Plot parameter values}
\arguments{
\item{obj}{rateModel}
}
\value{
a data.table with the parameter values and standard errors
}
\description{
Plots parameter values and standard errors
}
