% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\alias{plotTree,rateModel,rateModel-method}
\title{Tree visualization}
\arguments{
\item{model}{alleleData}

\item{offset}{How much to offset tip labels by}

\item{xmax}{Maximum value of x axis in tree plots (Note: if you want different x axis for different tree 
plots use \link[ggtree]{xlim_expand} on the returned object)}

\item{nodeLabels}{if TRUE label nodes with numbers}

\item{nodeLabels}{if TRUE label tips with names}
}
\description{
Plots tree with various helpful highlightings
}
\examples{

}
