% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict phylogenetic parameter values}
\arguments{
\item{object}{rateModel object}

\item{data}{a new set of genomic covariate values (siteInfo data)}

\item{rateDM}{allows direct setting of the design matrix for rate covariates 
(will override data, reccomended for advanced users only)}

\item{piDM}{allows direct setting of the design matrix for pi covariates 
(will override data, reccomended for advanced users only)}
}
\value{
a list containing both the rates and the stationary distribution at each site
}
\description{
Predicts the rate and stationary distribution per site
}
