% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateModel-class.R
\name{rateModel}
\alias{rateModel}
\title{rateModel}
\usage{
rateModel(data, rateFormula, piFormula = NULL, lineageTable = NULL)
}
\arguments{
\item{data}{An alleleData object}

\item{rateFormula}{A formula that uses the variables in the allele data object to compute turnover rate}

\item{piFormula}{A formula that uses the variables in the allele data object to compute pi (if NULL uses the same as rateFormula)}

\item{lineageTable}{A table with columns parent,child,edgeID, and rateGroup, where rateGroup is used to specify how the rates are tied between the branches}
}
\value{
an rateModel object
}
\description{
Contructs an object that holds a rate model for a set of alleles
}
\examples{

}
