% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rates}
\alias{rates}
\title{Compute rates}
\arguments{
\item{model}{rateModel object}

\item{sites}{integer vector of site indicies}
}
\value{
table of rates along with the corresponding edgeGroup and site
}
\description{
Compute rates of current model across specified sites
}
