% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPAF.R
\name{readPAF}
\alias{readPAF}
\title{Read in *.paf file}
\usage{
readPAF(f, subset)
}
\arguments{
\item{f}{path to *.paf file}

\item{subset}{vector of site labels to be included (defaults to all sites)}
}
\value{
a list of allele probability matricies and a list of the site labels for each row of the matrix
}
\description{
Read in *.paf file
}
