% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledLL.R
\name{scaledLL}
\alias{scaledLL}
\title{Compute scaled negative log-likelihood}
\arguments{
\item{x}{a set of parameters to compute the log-likelihood for (optional)}

\item{obj}{rateModel object}

\item{scale}{value to scale log-likelihood by (default=1)}

\item{threads}{number of threads to use}
}
\value{
scaled log-likelihood
}
\description{
Compute scaled log-likelihood for given rate model object
}
