% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\alias{se,rateModel,rateModel-method}
\title{Compute standard errors for parameters}
\arguments{
\item{obj}{rateModel}
}
\value{
a data.table with the parameter values and standard errors
}
\description{
Estimates the standard error for all non-fixed parameters in the model using the hessian
}
