% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFixed.R
\name{setFixed}
\alias{setFixed}
\title{Set fixed}
\arguments{
\item{obj}{rateModel object}

\item{index}{a 0-based integer vector of parameter indicies to be changed (note: parameters are 0 indexed)}

\item{x}{logical vector of whether parameter values are fixed}
}
\description{
Set whether parmeters aree fixed
}
