% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParams.R
\name{setParams}
\alias{setParams}
\alias{setParams,rateModel,rateModel-method}
\title{Set params}
\arguments{
\item{obj}{rateModel object}

\item{x}{parameter values}

\item{index}{an integer vector of parameter indicies to be changed (note: parameters are 0 indexed)}
}
\description{
Set parameter values for any length of parameters
}
