% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateBounds.R
\name{setRateBounds}
\alias{setRateBounds}
\alias{setRateBounds,rateModel,rateModel-method}
\title{Set rate bounds}
\arguments{
\item{model}{rateModel object}

\item{rMin}{rate lower bound}

\item{rMax}{rate upper bound}
}
\value{
vector of parameter values
}
\description{
Set rate bounds
}
