% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\title{Replace default print format}
\arguments{
\item{object}{alleleData or rateModel object}
}
\value{
A tree of class \code{phylo}
}
\description{
Method to replace default print appearance for
}
