% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSites.R
\name{simulateSites}
\alias{simulateSites}
\title{Simulate data}
\usage{
simulateSites(tr, covariateTable, rateFormula, rateParams = NULL,
  piFormula = NULL, piParams = NULL, lineageTable = NULL,
  rateBounds = c(10^-4, 5))
}
\arguments{
\item{tr}{tree}

\item{covariateTable}{a data.frame where each column is a covariate and each row
is a site (columns have covariate names)}

\item{rateFormula}{a formula object specifying how covariates are combined to compute the rate}

\item{rateParams}{a matrix of parameters where the columns correspond to the covariate label and the rows are
the edgeGroup from the lineage table}

\item{piFormula}{a formula object specifying how covariates are combined to compute the stationary distribution}

\item{piParams}{a matrix of parameters where the columns correspond to the covariate label and the rows are
each seperate alleles (note nAlleles = nrow(piParams)+1)}

\item{lineageTable}{a data.frame with three columns, parent, child, and edgeGroup. Can be used to specify different
rate parameters for different edges}

\item{rateBounds}{two element numeric vector which gives upper and lower bounds for rate}
}
\value{
list with the simulated data and the number of alleles
}
\description{
Simulates data without an error model
}
