% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteLL.R
\name{siteLL}
\alias{siteLL}
\alias{siteLL,rateModel,rateModel-method}
\title{Compute sitewise log-likelihoods}
\arguments{
\item{x}{rateModel object}

\item{sites}{integer vectore of sites that log-likelihoods should be computed for (note: 1-indexed)}
}
\value{
vector of parameter values
}
\description{
Compute sitewise log-likelihoods
}
