% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlMatrix-class.R
\name{stlMatrix-class}
\alias{stlMatrix-class}
\title{Class stlMatrix}
\description{
Class \code{stlMatrix} holds an external pointer to an stl matrix and had various methods
to make it easier to access elements
}
