% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlMatrix-class.R
\name{stlMatrix}
\alias{stlMatrix}
\title{stlMatrix}
\usage{
stlMatrix(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
an stlMatrix object
}
\description{
Contructs an object that holds a pointer to an stl matrix
}
