% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{unpack}
\alias{unpack}
\title{Unpack rateModel}
\arguments{
\item{pm}{a packedModel object}
}
\value{
a rateModel object
}
\description{
Unpacks a packedModel in to a rateModel
}
\examples{

}
