% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateFixed}
\alias{updateFixed}
\title{updateFixed}
\usage{
updateFixed(old, replacement, index)
}
\arguments{
\item{old}{A logical vector to be updated}

\item{replacement}{A logical vecotor of replacement values}

\item{index}{the 0-based indicies to update in the old vector}
}
\description{
This function takes an logical vector, a 0-based vector of indices, and an vector
of updated logical values to update the first logical vector without triggering
a rebuild of any associated S4 object.
}
