% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFormula.R
\name{updateFormula}
\alias{updateFormula}
\title{Update formula}
\arguments{
\item{model}{a rateModel object}

\item{rateFormula}{a new rate formula}

\item{piFormula}{a new pi formula}
}
\value{
a copy of the rateModel object with updated formulas (matching parameters carry over values)
}
\description{
Updates the rate and/or pi formula of a rate model
}
\examples{

}
