% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledLL.R
\name{ll}
\alias{ll}
\title{Compute log-likelihood}
\arguments{
\item{x}{a set of parameters to compute the log-likelihood for (optional)}

\item{model}{rateModel object}

\item{threads}{number of threads to use}

\item{index}{integer vector of parameter indicies in vector (0-based)}
}
\value{
scaled log-likelihood
}
\description{
Compute the log-likelihood for given rate model object.
}
