% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal}
\alias{marginal}
\title{Node marginals}
\arguments{
\item{obj}{rateModel object}

\item{threads}{number of threads to use}
}
\value{
a list of per node marginal posteriors
}
\description{
Compute the posterior marginal distribution across alleles per node
}
