% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginalTransitions}
\alias{marginalTransitions}
\title{Expected transitions per site}
\arguments{
\item{model}{rateModel object}

\item{aggregate}{How expected numbers of transitions should be aggregated (either by edge or by node)}

\item{threads}{number of threads to use}
}
\value{
a list of per node/edge matricies containing the expected number of allelic transitions. The 
rows are the ancestral allele and the columns are the descendant allele
}
\description{
Compute the expected number of transitions aggregated accorss either nodes or edges
}
