% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParams.R
\name{plotParams}
\alias{plotParams}
\alias{plotParams,rateModel,rateModel-method}
\alias{plotParams,data.frame,data.frame-method}
\title{Plot parameter values}
\arguments{
\item{x}{rateModel object or the data.frame from se(rateModel)}

\item{alpha}{plot the confidence interval of the parameters from [alpha/2,1-alpha/2]}
}
\value{
a data.table with the parameter values and standard errors
}
\description{
Plots parameter values and standard errors
}
