% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\alias{se,rateModel,rateModel-method}
\title{Compute standard errors for parameters}
\arguments{
\item{x}{rateModel object}

\item{hess}{hessian matrix for (not supplied if x is a rateModel object)}
}
\value{
a data.table with the parameter values and standard errors
}
\description{
Estimates the standard error for all non-fixed parameters in the model using the hessian
}
