% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteLL.R
\name{siteLL}
\alias{siteLL}
\alias{siteLL,rateModel,rateModel-method}
\title{Compute sitewise log-likelihoods}
\arguments{
\item{model}{rateModel object}

\item{threads}{number of threads to use}
}
\value{
vector of parameter values
}
\description{
Compute sitewise log-likelihoods
}
